# sistema-ventas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Ventas",
        "description": "Documentación de la API para el sistema de ventas",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener todos los productos",
                "description": "Devuelve una lista de todos los productos registrados en el sistema",
                "operationId": "getAllProducts",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "PROD001"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Producto de ejemplo"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Descripción del producto"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10.99
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "product_category_id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto en el sistema",
                "operationId": "createProduct",
                "requestBody": {
                    "description": "Datos del producto a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "stock",
                                    "product_category_id"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "PROD001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Producto de ejemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del producto"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "product_category_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "550e8400-e29b-41d4-a716-446655440001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440006"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "PROD001"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Producto de ejemplo"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción del producto"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 10.99
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "product_category_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440001"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El código del producto no puede estar vacío."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear el producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/category/{productCategoryId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener productos por categoría",
                "description": "Devuelve una lista de productos que pertenecen a una categoría específica",
                "operationId": "getProductsByCategory",
                "parameters": [
                    {
                        "name": "productCategoryId",
                        "in": "path",
                        "description": "ID de la categoría de productos",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos de la categoría",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "PROD001"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Producto de ejemplo"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Descripción del producto"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 10.99
                                            },
                                            "stock": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "product_category_id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": "550e8400-e29b-41d4-a716-446655440000"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}